[patch] readmes: Add CONTRIBUTIONS notice
The current project lacks a coherent way to add new features and manage
bugs. Potential contributors do not have a standard to follow to keep their
commit messages clear and clean.

This commit introduces a README update, and a new CONTRIBUTING file, which
outline the project, and describe how one can offer new code and features.
These are still drafts, and not very helpful yet.